+++ src/resources/lang/pt_br/permissionmanager.php
@@ -1,0 +1,30 @@
+<?php
+
+return [
+    /*
+    |--------------------------------------------------------------------------
+    | Permission Manager Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are used for Laravel Backpack - Permission Manager
+    | Author: Lúdio Oliveira <ludio.ao@gmail.com>
+    |
+    */
+
+    'name'                  => 'Nome',
+    'role'                  => 'Grupo de Usuário',
+    'roles'                 => 'Grupos de Usuário',
+    'roles_have_permission' => 'Grupos que possuem esta permissão',
+    'permission_singular'   => 'permissão',
+    'permission_plural'     => 'permissões',
+    'user_singular'         => 'Usuário',
+    'user_plural'           => 'Usuários',
+    'email'                 => 'Email',
+    'extra_permissions'     => 'Permissões Extras',
+    'password'              => 'Senha',
+    'password_confirmation' => 'Confirmar senha',
+    'user_role_permission'  => 'Permissões do Grupo de Usuário',
+    'user'                  => 'Usuário',
+    'users'                 => 'Usuários',
+
+];